refactor: 중복 로직 제거
